Refactor initializers and move models
