+++ src/app/components/confirmation-dialog/confirmation-dialog.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
-import { DialogData } from '../../shared/models/dialog-data.model';
+import { DialogData } from '../../models/dialog-data.model';
 
 @Component({
   templateUrl: './confirmation-dialog.component.html',
