+++ src/app/core/auth/auth.providers.ts
@@ -3,14 +3,14 @@ import { EnvironmentProviders, Provider } from '@angular/core';
 import { PreloadAllModules, withPreloading } from '@angular/router';
 import { provideRootRouter } from '../../common/router-features';
 import { AUTH_ROUTES } from './auth.routes';
-import { INIT_USER_CONTEXT_PROVIDER } from './initializers/auth.initializer';
+import { AUTH_INITIALIZER_PROVIDERS } from './initializers/auth.initializer';
 import { authExpiredInterceptor } from './interceptors/auth-expired.interceptor';
 
 export const AUTH_PROVIDERS: (Provider | EnvironmentProviders)[] = [
   provideHttpClient(withInterceptors([
     authExpiredInterceptor(),
   ])),
-  INIT_USER_CONTEXT_PROVIDER,
+  AUTH_INITIALIZER_PROVIDERS,
   provideRootRouter(AUTH_ROUTES,
     withPreloading(PreloadAllModules),
   ),
