+++ src/app/pages/home/groups/charge/components/charge-mode/charge-mode.component.ts
@@ -6,8 +6,8 @@ import { MatIconModule } from '@angular/material/icon';
 import { NgxKamereonClient } from '@remscodes/ngx-renault-api-client';
 import { ChargingSettings, DateType, Day } from '@remscodes/renault-api';
 import { finalize } from 'rxjs';
+import { DAYS } from '../../../../../../core/renault/constants/day.constants';
 import { fade } from '../../../../../../shared/animations/fade.animation';
-import { DAYS } from '../../../../../../shared/contants/day.constants';
 import { BetterRouter } from '../../../../../../shared/services/better-router.service';
 import { Loading } from '../../../../../../shared/services/loading.service';
 import { AddTimeScheduleDialogComponent } from '../../dialogs/add-time-schedule-dialog/add-time-schedule-dialog.component';
